Decoupled Systems and Scenes. Every EngineState now owns a separate instance of a SystemScene and all the Systems are shared and initialized once Moved Systems initialization from EngineState to Engine Added scene arrays in every System, methods for creating, getting and destroying SystemScenes Added a way to reorder RenderPasses in the same RenderSystem instance. RenderScene now owns a list of RenderPass orders Added initialization flags to every RenderPass Added more error codes and error strings Fixed a bug of GUI Docking windows being left enabled after exiting EditorState, thus rendering over the whole screen in PlayState Fixed a bug in PhysicsScene of heap corruption during the scene destruction

+++ Praxis3D/Source/WorldScene.h
@@ -162,7 +162,7 @@ public:
 
 	// Get the internal pool size of a given component
 	template <class T_Component>
-	unsigned int getPoolSize()
+	size_t getPoolSize()
 	{
 		return m_entityRegistry.storage<T_Component>().size();
 	}
